Fix radar rendering within the Editor (#8853)
