+++ src/fheroes2/editor/editor_interface.cpp
@@ -401,33 +401,6 @@ namespace
         return needRedraw;
     }
 
-    void updateWorldCastlesHeroes( const Maps::Map_Format::MapFormat & map )
-    {
-        const auto & townObjects = Maps::getObjectsByGroup( Maps::ObjectGroup::KINGDOM_TOWNS );
-        const auto & heroObjects = Maps::getObjectsByGroup( Maps::ObjectGroup::KINGDOM_HEROES );
-
-        for ( size_t i = 0; i < map.tiles.size(); ++i ) {
-            for ( const auto & object : map.tiles[i].objects ) {
-                if ( object.group == Maps::ObjectGroup::KINGDOM_TOWNS ) {
-                    const uint8_t color = Color::IndexToColor( Maps::getTownColorIndex( map, i, object.id ) );
-                    const uint8_t race = Race::IndexToRace( static_cast<int>( townObjects[object.index].metadata[0] ) );
-
-                    world.addCastle( static_cast<int32_t>( i ), race, color );
-                }
-                else if ( object.group == Maps::ObjectGroup::KINGDOM_HEROES ) {
-                    const auto & metadata = heroObjects[object.index].metadata;
-                    const uint8_t color = Color::IndexToColor( static_cast<int>( metadata[0] ) );
-
-                    Heroes * hero = world.GetHeroForHire( static_cast<int>( metadata[1] ) );
-                    if ( hero ) {
-                        hero->SetCenter( { static_cast<int32_t>( i ) % world.w(), static_cast<int32_t>( i ) / world.w() } );
-                        hero->SetColor( color );
-                    }
-                }
-            }
-        }
-    }
-
     bool verifyTerrainPlacement( const fheroes2::Point & tilePos, const Maps::ObjectGroup groupType, const int32_t objectType, std::string & errorMessage )
     {
         switch ( groupType ) {
@@ -962,8 +935,6 @@ namespace Interface
 
                             // TODO: Make a proper function to remove all types of objects from the 'world tiles' not to do full reload of '_mapFormat'.
                             Maps::readMapInEditor( _mapFormat );
-
-                            updateWorldCastlesHeroes( _mapFormat );
                         }
                     }
                 }
@@ -1356,8 +1327,6 @@ namespace Interface
 
                 // TODO: Make a proper function to remove all types of objects from the 'world tiles' not to do full reload of '_mapFormat'.
                 Maps::readMapInEditor( _mapFormat );
-
-                updateWorldCastlesHeroes( _mapFormat );
             }
 
             if ( brushSize.width == 0 ) {
@@ -1614,29 +1583,6 @@ namespace Interface
         Editor::showPopupWindow( world.GetTiles( tileIndex ) );
     }
 
-    void EditorInterface::mouseCursorAreaLongPressLeft( const int32_t /*Unused*/ )
-    {
-        // Do nothing.
-    }
-
-    void EditorInterface::undoAction()
-    {
-        if ( _historyManager.undo() ) {
-            _redraw |= ( REDRAW_GAMEAREA | REDRAW_RADAR );
-
-            updateWorldCastlesHeroes( _mapFormat );
-        }
-    }
-
-    void EditorInterface::redoAction()
-    {
-        if ( _historyManager.redo() ) {
-            _redraw |= ( REDRAW_GAMEAREA | REDRAW_RADAR );
-
-            updateWorldCastlesHeroes( _mapFormat );
-        }
-    }
-
     void EditorInterface::updateCursor( const int32_t tileIndex )
     {
         if ( _cursorUpdater && tileIndex >= 0 ) {
@@ -1690,8 +1636,6 @@ namespace Interface
             fheroes2::showStandardTextMessage( _( "Warning!" ), "Failed to read the map.", Dialog::OK );
             return false;
         }
-
-        updateWorldCastlesHeroes( _mapFormat );
 
         _loadedFileName = System::truncateFileExtensionAndPath( filePath );
