+++ src/fheroes2/editor/editor_interface.h
@@ -66,10 +66,25 @@ namespace Interface
 
         void mouseCursorAreaClickLeft( const int32_t tileIndex ) override;
         void mouseCursorAreaPressRight( const int32_t tileIndex ) const override;
-        void mouseCursorAreaLongPressLeft( const int32_t /*Unused*/ ) override;
 
-        void undoAction();
-        void redoAction();
+        void mouseCursorAreaLongPressLeft( const int32_t /*Unused*/ ) override
+        {
+            // Do nothing.
+        }
+
+        void undoAction()
+        {
+            if ( _historyManager.undo() ) {
+                _redraw |= ( REDRAW_GAMEAREA | REDRAW_RADAR );
+            }
+        }
+
+        void redoAction()
+        {
+            if ( _historyManager.redo() ) {
+                _redraw |= ( REDRAW_GAMEAREA | REDRAW_RADAR );
+            }
+        }
 
         void updateCursor( const int32_t tileIndex ) override;
